move credential provider interface to credentials package
