+++ internal/vault/cache_metrics_test.go
@@ -15,7 +15,7 @@ import (
 	"k8s.io/apimachinery/pkg/types"
 
 	secretsv1alpha1 "github.com/hashicorp/vault-secrets-operator/api/v1alpha1"
-	"github.com/hashicorp/vault-secrets-operator/internal/vault/credentialproviders"
+	"github.com/hashicorp/vault-secrets-operator/internal/vault/credentials"
 )
 
 func Test_clientCacheCollector_Collect(t *testing.T) {
@@ -61,7 +61,7 @@ func Test_clientCacheCollector_Collect(t *testing.T) {
 							Generation: 0,
 						},
 					},
-					credentialProvider: credentialproviders.NewKubernetesCredentialProvider(nil, "",
+					credentialProvider: credentials.NewKubernetesCredentialProvider(nil, "",
 						types.UID(uuid.New().String())),
 				})
 				require.NoError(t, err)
@@ -173,7 +173,7 @@ func Test_clientCache_Metrics(t *testing.T) {
 							Generation: 0,
 						},
 					},
-					credentialProvider: credentialproviders.NewKubernetesCredentialProvider(nil, "",
+					credentialProvider: credentials.NewKubernetesCredentialProvider(nil, "",
 						types.UID(uuid.New().String())),
 				}
