+++ internal/vault/cache_storage_metrics_test.go
@@ -23,7 +23,7 @@ import (
 
 	secretsv1alpha1 "github.com/hashicorp/vault-secrets-operator/api/v1alpha1"
 	"github.com/hashicorp/vault-secrets-operator/internal/metrics"
-	"github.com/hashicorp/vault-secrets-operator/internal/vault/credentialproviders"
+	"github.com/hashicorp/vault-secrets-operator/internal/vault/credentials"
 )
 
 func Test_clientCacheStorageCollector(t *testing.T) {
@@ -534,7 +534,7 @@ func storeSecret(t *testing.T, ctx context.Context, client ctrlclient.Client, st
 					Generation: 0,
 				},
 			},
-			credentialProvider: credentialproviders.NewKubernetesCredentialProvider(nil, "",
+			credentialProvider: credentials.NewKubernetesCredentialProvider(nil, "",
 				types.UID(uuid.New().String())),
 		},
 	}
