+++ internal/vault/credentials/consts.go
@@ -1,7 +1,7 @@
 // Copyright (c) HashiCorp, Inc.
 // SPDX-License-Identifier: MPL-2.0
 
-package credentialproviders
+package credentials
 
 const (
 	TokenGenerateName string = "vso-"
