+++ internal/vault/credentials/credential_provider.go
@@ -1,7 +1,7 @@
 // Copyright (c) HashiCorp, Inc.
 // SPDX-License-Identifier: MPL-2.0
 
-package vault
+package credentials
 
 import (
 	"context"
@@ -11,7 +11,6 @@ import (
 	ctrlclient "sigs.k8s.io/controller-runtime/pkg/client"
 
 	secretsv1alpha1 "github.com/hashicorp/vault-secrets-operator/api/v1alpha1"
-	"github.com/hashicorp/vault-secrets-operator/internal/vault/credentialproviders"
 )
 
 const (
@@ -19,12 +18,7 @@ const (
 	providerMethodJWT        string = "jwt"
 )
 
-var (
-	providerMethodsSupported = []string{providerMethodKubernetes, providerMethodJWT}
-
-	_ CredentialProvider = (*credentialproviders.KubernetesCredentialProvider)(nil)
-	_ CredentialProvider = (*credentialproviders.JWTCredentialProvider)(nil)
-)
+var providerMethodsSupported = []string{providerMethodKubernetes, providerMethodJWT}
 
 type CredentialProvider interface {
 	Init(ctx context.Context, client ctrlclient.Client, object *secretsv1alpha1.VaultAuth, providerNamespace string) error
@@ -40,13 +34,13 @@ func NewCredentialProvider(ctx context.Context, client ctrlclient.Client, authOb
 
 	switch authObj.Spec.Method {
 	case providerMethodJWT:
-		provider := &credentialproviders.JWTCredentialProvider{}
+		provider := &JWTCredentialProvider{}
 		if err := provider.Init(ctx, client, authObj, providerNamespace); err != nil {
 			return nil, err
 		}
 		return provider, nil
 	case providerMethodKubernetes:
-		provider := &credentialproviders.KubernetesCredentialProvider{}
+		provider := &KubernetesCredentialProvider{}
 		if err := provider.Init(ctx, client, authObj, providerNamespace); err != nil {
 			return nil, err
 		}
